Make ok! cleaning. no cpp except show_program.f90. All real(8). Remove unused .f90

+++ SRC/subroutines/cpudel.f90
@@ -47,7 +47,7 @@ real(8) function cpusec()
   logical,save::firsttime=.true.
   integer,save::i1
   integer:: i2,irate,imax
-  real(4)::diff
+  real(8)::diff
 
   if (firsttime) then
      call system_clock(i1)
